requirements.txt
# Legal AI Service - Dependencies
# Generated from project files analysis

# Flask and SQLAlchemy (models.py)
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
Werkzeug==3.0.1

# Document generation (document_generator.py)
python-docx==1.1.0

# Authentication (auth.py)
PyJWT==2.8.0

# API calls (kimi_api.py)
requests==2.31.0

# Common utilities
python-dotenv==1.0.0

# Database (SQLite support built-in, but adding for clarity)
# SQLite3 is part of Python standard library
